refactor: 인증 관련 유틸리티 함수 통합
